requirements.txt
python-dotenv~=1.1.0
selenium~=4.32.0
google-api-python-client~=2.169.0
beautifulsoup4~=4.13.4
webdriver-manager~=4.0.2
yfinance~=0.2.61
colorama~=0.4.6
protobuf~=6.31.0
requests~=2.32.3
PyPDF2~=3.0.1